moved code to __init__
